+++ hassio-addon/run.sh
@@ -32,6 +32,9 @@ get_mqtt_uri() {
 # Create config directory
 mkdir -p /app/config
 
+# Copy default brokers configuration
+cp /app/brokers.json /app/config/brokers.json
+
 # Get MQTT URI
 BROKER_URL=$(get_mqtt_uri)
 
@@ -39,6 +42,7 @@ BROKER_URL=$(get_mqtt_uri)
 bashio::log.info "Generating config file..."
 DEVICES=$(bashio::config 'devices' | jq -s '.')
 INVERSE_FORWARDING=$(bashio::config 'inverse_forwarding' "false")
+DEFAULT_BROKER_ID=$(bashio::config 'default_broker_id' "hame-2024")
 
 # Create base configuration
 CONFIG='{
@@ -62,13 +66,15 @@ if bashio::config.has_value 'username'; then
           --arg url "$BROKER_URL" \
           --argjson devices "$DEVICES" \
           --argjson inverse "$INVERSE_FORWARDING" \
+          --arg default "$DEFAULT_BROKER_ID" \
           --arg username "$USERNAME" \
           --arg password "$PASSWORD" \
           '{
-            broker_url: $url, 
-            devices: $devices, 
-            inverse_forwarding: $inverse, 
-            username: $username, 
+            broker_url: $url,
+            devices: $devices,
+            inverse_forwarding: $inverse,
+            default_broker_id: $default,
+            username: $username,
             password: $password
           }' > /app/config/config.json
     else
@@ -77,11 +83,13 @@ if bashio::config.has_value 'username'; then
           --arg url "$BROKER_URL" \
           --argjson devices "$DEVICES" \
           --argjson inverse "$INVERSE_FORWARDING" \
+          --arg default "$DEFAULT_BROKER_ID" \
           --arg username "$USERNAME" \
           '{
-            broker_url: $url, 
-            devices: $devices, 
-            inverse_forwarding: $inverse, 
+            broker_url: $url,
+            devices: $devices,
+            inverse_forwarding: $inverse,
+            default_broker_id: $default,
             username: $username
           }' > /app/config/config.json
     fi
@@ -91,10 +99,12 @@ else
       --arg url "$BROKER_URL" \
       --argjson devices "$DEVICES" \
       --argjson inverse "$INVERSE_FORWARDING" \
+      --arg default "$DEFAULT_BROKER_ID" \
       '{
-        broker_url: $url, 
-        devices: $devices, 
-        inverse_forwarding: $inverse
+        broker_url: $url,
+        devices: $devices,
+        inverse_forwarding: $inverse,
+        default_broker_id: $default
       }' > /app/config/config.json
 fi